requirements.txt
pytest==8.2.1
allure-pytest==2.13.5